split validator for withdrawal gating vs withdrawal capacity
